feat: change Double type to BigDecimal for avg and percentile

+++ src/main/java/backend/academy/loganalyzer/models/Report.java
@@ -2,6 +2,7 @@ package backend.academy.loganalyzer.models;
 
 import backend.academy.loganalyzer.enums.Filter;
 import backend.academy.loganalyzer.enums.Format;
+import java.math.BigDecimal;
 import java.nio.file.Path;
 import java.time.LocalDate;
 import java.util.List;
@@ -15,8 +16,8 @@ public record Report(
     Filter filter,
     String filterValue,
     Integer requestCount,
-    Double avgSize,
-    Double percentile,
+    BigDecimal avgSize,
+    BigDecimal percentile,
     Map<String, Integer> resources,
     Map<Integer, Integer> popularCodes
 ) {
